# be-summit/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Summit API Documentation",
        "description": "API documentation for the BE Summit project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://be-summit.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/kyc": {
            "get": {
                "tags": [
                    "KYC (Admin)"
                ],
                "summary": "List all climber KYC profiles (Admin)",
                "operationId": "6d7ae9ea3562503c9c3ea2451ce8dfce",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by verification status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "disetujui",
                                "ditolak"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KYC profiles fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "KYC profiles fetched successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PendakiResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - only for admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/kyc/{id}": {
            "get": {
                "tags": [
                    "KYC (Admin)"
                ],
                "summary": "Get details of a specific KYC profile (Admin)",
                "operationId": "8e55726d3bf024765be9334edfdef054",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "KYC / Pendaki Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KYC profile fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "KYC profile details fetched successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PendakiResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - only for admin"
                    },
                    "404": {
                        "description": "KYC profile not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/kyc/{id}/download-document": {
            "get": {
                "tags": [
                    "KYC (Admin)"
                ],
                "summary": "Download climber uploaded identity document file securely (Admin)",
                "operationId": "0e57dc78838e2fda27da3f81a04e33f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "KYC / Pendaki Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secure file download response",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - only for admin"
                    },
                    "404": {
                        "description": "KYC profile or document file not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/kyc/{id}/verify": {
            "post": {
                "tags": [
                    "KYC (Admin)"
                ],
                "summary": "Verify (approve or reject) a climber KYC profile (Admin)",
                "operationId": "91660f81dbbb869fd0a037aa8c8a51e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "KYC / Pendaki Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_verifikasi"
                                ],
                                "properties": {
                                    "status_verifikasi": {
                                        "type": "string",
                                        "example": "disetujui",
                                        "enum": [
                                            "disetujui",
                                            "ditolak"
                                        ]
                                    },
                                    "alasan_penolakan": {
                                        "type": "string",
                                        "example": "Foto KTP buram dan tidak terbaca."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "KYC profile verified successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "KYC profile verified successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PendakiResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - only for admin"
                    },
                    "404": {
                        "description": "KYC profile not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PendakiResource": {
                "title": "Pendaki Resource",
                "description": "Climber profile KYC resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_lengkap": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "jenis_identitas": {
                        "type": "string",
                        "example": "ktp"
                    },
                    "nomor_identitas": {
                        "type": "string",
                        "example": "1234567890123456"
                    },
                    "foto_identitas": {
                        "type": "string",
                        "example": "kyc_documents/xyz.jpg"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "jenis_kelamin": {
                        "type": "string",
                        "example": "l"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 45"
                    },
                    "telepon": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "nama_kontak_darurat": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "telepon_darurat": {
                        "type": "string",
                        "example": "081298765432"
                    },
                    "hubungan_darurat": {
                        "type": "string",
                        "example": "Istri"
                    },
                    "status_verifikasi": {
                        "type": "string",
                        "example": "pending"
                    },
                    "alasan_penolakan": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "verified_by": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}